requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.6
django-admin-rangefilter==0.13.2
django-autocomplete-light==3.11.0
django-import-export==4.3.5
django-rangefilter==0.9.1
sqlparse==0.5.3
tablib==3.8.0
